ci: fix security scan workflow (#173)
* fix workflow trigger

The workflow actually needs to run on **push** event with branch and path
filtering rules applied and not pull_request. This approach ensures that
the workflow runs only when needed - updates to `Dockerfile` in version
bump pull requests.

* fix paths used by trivy

By default repository code is checked out to current working directory
and that's where trivy needs to look for its configuration file as well as
scan ref when running in fs mode.